requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
firebase-admin
python-dotenv
gunicorn
